requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.6
python-multipart==0.0.17
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
openai==1.61.1
httpx==0.28.1
